Head compiler warnings.
git-svn-id: https://anonsvn.ncbi.nlm.nih.gov/repos/v1/trunk/c++@52889 78c7ea69-d796-4a43-9a09-de51944f1b03

+++ src/dbapi/test/dbapi_unit_test_lob.cpp
@@ -1077,7 +1077,7 @@ BOOST_AUTO_TEST_CASE(Test_BlobStream)
                 out << i << " ";
             }
 
-            data_len = out.pcount();
+            data_len = long(out.pcount());
             BOOST_CHECK(data_len > 0);
 
             // Clean table ...
@@ -1152,7 +1152,7 @@ BOOST_AUTO_TEST_CASE(Test_BlobStream)
                                 BOOST_CHECK_EQUAL(strm.eof(), false);
                                 BOOST_CHECK_EQUAL(j, i);
                             }
-                            long read_data_len = strm.tellg();
+                            long read_data_len = long(strm.tellg());
                             // Calculate a trailing space.
                             BOOST_CHECK_EQUAL(data_len, read_data_len + 1);
                         }
